Product Admin - Seuring the Views

+++ products/views.py
@@ -1,6 +1,7 @@
 # pylint: disable=no-member
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -91,8 +92,15 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+# login_required checks if the user is logged in
+# otherwise it will redirect the user to the login page
+@login_required
 def add_product(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry only the store owners can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         """
         Add form fields and files from image url
@@ -119,8 +127,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """Edit a product in the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry only the store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -146,8 +159,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """Delete a prduct from the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry only the store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
